+++ Scripts/Map/Domain/ITerrainData.cs
@@ -11,6 +11,6 @@ namespace SciencePotato.Scripts.Map.Domain
 	{
 		string Id { get; set; }
 		string Name { get; set; }
-		double Weight { get; set; }
+		float Weight { get; set; }
 	}
 }
