+++ Scripts/Map/Infrastructure/TerrainConfigResources.cs
@@ -13,6 +13,6 @@ namespace SciencePotato.Scripts.Map.Infrastructure
 	{
 		[Export] public string Id { get; set; }
 		[Export] public string Name { get; set; }
-		[Export] public double Weight { get; set; }
+		[Export] public float Weight { get; set; }
 	}
 }
