+++ Scripts/Map/Presentation/MapCellView.cs
@@ -37,6 +37,6 @@ public partial class MapCellView : Node2D
         // Set to the right Position with correct seperation
         int q = CellPosition.ToCoordinate().Item1;
         int r = CellPosition.ToCoordinate().Item2;
-        base.Position = new Vector2(height / 2 * r - height * (float)Math.Ceiling((double)r*0.5d)+ height * q,r * width*3/4);
+        base.Position = new Vector2(height / 2 * r - height * (float)Math.Ceiling((float)r*0.5d)+ height * q,r * width*3/4);
 	}
 }
